src/central/requirements.txt
together==1.5.8
fastapi==0.115.12
requests==2.32.3
uvicorn==0.34.2
pyyaml==6.0.2